more fixes

+++ tests/speed_test.h
@@ -88,7 +88,7 @@ public:
     	typedef Particles<std::tuple<a_var,b_var,c_var>,3> nodes_type;
        	nodes_type nodes(N);
         //<- 
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             get<a_var>(nodes)[i] = i;
             get<b_var>(nodes)[i] = i*2;
         }
@@ -125,7 +125,7 @@ public:
         ABORIA_VARIABLE(c_var,double,"c")
     	typedef Particles<std::tuple<a_var,b_var,c_var>,3> nodes_type;
        	nodes_type nodes(N);
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             get<a_var>(nodes)[i] = i;
             get<b_var>(nodes)[i] = i*2;
         }
@@ -134,13 +134,13 @@ public:
         We compare this with Level 1 Aboria using the `get` functions and looping
         through the container
         */
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             get<a_var>(nodes)[i] = get<b_var>(nodes)[i] + get<c_var>(nodes)[i];
         }
         //<-
         auto t0 = Clock::now();
         for (int r=0; r<repeats; ++r) {
-            for (int i=0; i<N; i++) {
+            for (size_t i=0; i<N; i++) {
                 get<a_var>(nodes)[i] = get<b_var>(nodes)[i] + get<c_var>(nodes)[i];
             }
         }
@@ -157,18 +157,18 @@ public:
         */
         std::vector<double> a(N),b(N),c(N);
         //<-
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             a[i] = i;
             b[i] = i*2;
         }
         //->
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             a[i] = b[i] + c[i];
         }
         //<-
         auto t0 = Clock::now();
         for (int r=0; r<repeats; ++r) {
-            for (int i=0; i<N; i++) {
+            for (size_t i=0; i<N; i++) {
                 a[i] = b[i] + c[i];
             }
         }
@@ -187,7 +187,7 @@ public:
         typedef Eigen::Matrix<double,Eigen::Dynamic,1> vector_type; 
         vector_type a(N),b(N),c(N);
         //<-
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             a[i] = i;
             b[i] = 2*i;
         }
@@ -222,7 +222,7 @@ public:
         ABORIA_VARIABLE(b_var,double,"b")
     	typedef Particles<std::tuple<a_var,b_var>,3> nodes_type;
        	nodes_type nodes(N);
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             get<a_var>(nodes)[i] = i;
             get<b_var>(nodes)[i] = i*2;
         }
@@ -256,7 +256,7 @@ public:
         ABORIA_VARIABLE(b_var,double,"b")
     	typedef Particles<std::tuple<a_var,b_var>,3> nodes_type;
        	nodes_type nodes(N);
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             get<a_var>(nodes)[i] = i;
             get<b_var>(nodes)[i] = i*2;
         }
@@ -264,13 +264,13 @@ public:
         /*`
         We compare against a Level 1 implementation like so 
         */
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             get<a_var>(nodes)[i] += 0.1*get<b_var>(nodes)[i];
         }
         //<-
         auto t0 = Clock::now();
         for (int r=0; r<repeats; ++r) {
-            for (int i=0; i<N; i++) {
+            for (size_t i=0; i<N; i++) {
                 get<a_var>(nodes)[i] += 0.1*get<b_var>(nodes)[i];
             }
         }
@@ -287,18 +287,18 @@ public:
         */
         std::vector<double> a(N),b(N);
         //<-
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             a[i] = i;
             b[i] = i*2;
         }
         //->
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             a[i] += 0.1*b[i];
         }
         //<-
         auto t0 = Clock::now();
         for (int r=0; r<repeats; ++r) {
-            for (int i=0; i<N; i++) {
+            for (size_t i=0; i<N; i++) {
                 a[i] += 0.1*b[i];
             }
         }
@@ -318,7 +318,7 @@ public:
         typedef Eigen::Matrix<double,Eigen::Dynamic,1> vector_type; 
         vector_type a(N),b(N);
         //<-
-        for (int i=0; i<N; i++) {
+        for (size_t i=0; i<N; i++) {
             a[i] = i;
             b[i] = 2*i;
         }
